add,delete,and resend feature done

+++ apollo-client/src/actions/index.js
@@ -1,5 +1,7 @@
 import ApolloClient from 'apollo-boost';
 import gql from 'graphql-tag';
+import Swal from 'sweetalert2'
+
 
 const API_URL = 'http://localhost:3001/graphql/'
 
@@ -31,13 +33,13 @@ export const loadPhone = () => {
     return client.query({
       query: usersQuery,
     })
-    .then(function (response) {
-      dispatch(loadPhoneSuccess(response.data.phones))
-    })
-    .catch(function (error) {
-      console.error(error);
-      dispatch(loadPhoneFailure())
-    });
+      .then(function (response) {
+        dispatch(loadPhoneSuccess(response.data.phones))
+      })
+      .catch(function (error) {
+        console.error(error);
+        dispatch(loadPhoneFailure())
+      });
   }
 }
 
@@ -50,16 +52,16 @@ export const postPhoneSuccess = (phones) => ({
   phones
 })
 
-export const postPhoneFailure = (PhoneNumber) => ({
-  type: 'POST_PHONE_FAILURE', PhoneNumber
+export const postPhoneFailure = (id) => ({
+  type: 'POST_PHONE_FAILURE', id
 })
 
-const postPhoneRedux = (PhoneNumber, Name,id) => ({
-  type: 'POST_PHONE', PhoneNumber, Name,id
+const postPhoneRedux = (PhoneNumber, Name, id) => ({
+  type: 'POST_PHONE', PhoneNumber, Name, id
 })
 
 
-export const postPhone = (PhoneNumber, Name,id) => {
+export const postPhone = (PhoneNumber, Name, id) => {
   const addQuery = gql`
   mutation addContact($Name: String!, $PhoneNumber: String!,$id:ID!) {
     addContact(Name: $Name, PhoneNumber: $PhoneNumber,id:$id) {
@@ -68,22 +70,41 @@ export const postPhone = (PhoneNumber, Name,id) => {
     }
   }`;
   return dispatch => {
-    dispatch(postPhoneRedux(PhoneNumber, Name,id))
-    return client.mutate({
-      mutation: addQuery,
-      variables: {
-        PhoneNumber,
-        Name,
-        id
-      }
+    Swal.fire({
+      position: 'center',
+      icon: 'success',
+      title: 'Contact added successfully!',
+      showConfirmButton: false,
+      timer: 1200
+    }).then(() => {
+      dispatch(postPhoneRedux(PhoneNumber, Name, id))
+      return client.mutate({
+        mutation: addQuery,
+        variables: {
+          PhoneNumber,
+          Name,
+          id
+        }
+      })
+        .then(function (response) {
+          dispatch(postPhoneSuccess(response.data))
+        })
+        .catch(function (error) {
+          Swal.fire({
+            icon: 'warning',
+            title: "Network connection trouble!",
+            text: "Click resend button to add your data!",
+            type: "warning",
+            buttons: true,
+            dangerMode: true,
+            timer: 1500
+          }).then(() => {
+            dispatch(postPhoneFailure(id))
+          })
+
+        });
     })
-    .then(function (response) {
-      dispatch(postPhoneSuccess(response.data))
-    })
-    .catch(function (error) {
-      console.error(error);
-      dispatch(postPhoneFailure(PhoneNumber))
-    });
+
   }
 }
 
@@ -112,20 +133,37 @@ export const deletePhone = (id) => {
     }
   }`;
   return dispatch => {
-    dispatch(deletePhoneRedux(id))
-    return client.mutate({
-      mutation: deleteQuery,
-      variables: {
-        id
+    Swal.fire({
+      icon: 'warning',
+      title: "Are you sure delete this contact?",
+      text: "You can't revert this action",
+      type: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes Delete it!",
+      cancelButtonText: "No, Keep it!",
+      showCloseButton: true,
+      showLoaderOnConfirm: true
+    }).then(result => {
+      console.log('test delete')
+      console.log(result)
+      if (result.value) {
+        dispatch(deletePhoneRedux(id))
+        return client.mutate({
+          mutation: deleteQuery,
+          variables: {
+            id
+          }
+        })
+          .then(function (response) {
+            dispatch(deletePhoneSuccess(response))
+          })
+          .catch(function (error) {
+            console.error(error);
+            dispatch(deletePhoneFailure())
+          });
       }
     })
-    .then(function (response) {
-      dispatch(deletePhoneSuccess(response))
-    })
-    .catch(function (error) {
-      console.error(error);
-      dispatch(deletePhoneFailure())
-    });
+
   }
 }
 
@@ -138,10 +176,10 @@ const resendPhoneSuccess = (id) => ({
 })
 
 
-export const resendPhone = (PhoneNumber, Name,id) => {
+export const resendPhone = (PhoneNumber, Name, id) => {
   const addQuery = gql`
-  mutation addContact($PhoneNumber: String!, $Name: String!) {
-    addContact(PhoneNumber: $PhoneNumber, Name: $Name) {
+  mutation addContact($PhoneNumber: String!, $Name: String!,$id:ID!) {
+    addContact(PhoneNumber: $PhoneNumber, Name: $Name,id:$id) {
       PhoneNumber
       Name
     }
@@ -151,17 +189,26 @@ export const resendPhone = (PhoneNumber, Name,id) => {
       mutation: addQuery,
       variables: {
         PhoneNumber,
-        Name
+        Name,
+        id
       }
     })
-    .then(function (response) {
-      dispatch(resendPhoneSuccess(id))
-    })
-    .catch(function (error) {
-      console.error(error);
-      dispatch(postPhoneFailure(id))
-    });
+      .then(function (response) {
+        dispatch(resendPhoneSuccess(id))
+      })
+      .catch(function (error) {
+        console.error(error);
+        dispatch(postPhoneFailure(id))
+      });
   }
 }
+const togleThisButton = () => ({
+  type: 'TOGLE'
+})
 
 
+export const TogleButtonCta=()=>{
+  return dispatch=>{
+    dispatch(togleThisButton())
+  }
+}
